Support split packages
Split packages are now supported. The "name" field of the pkgbuild_t
structure was changed to "names" and is now an array of names. These
elements correspond to the names of split packages. A "basename" field
was added, which refers to the group of packages (see "pkgbase" in man 5
PKGBUILD). The new "splitpkgs" field is an array of pkgbuild_t
structures. Technically this should not be the case, as split packages
can only override a subset of PKGBUILD fields, however this was done for
simplicity and to reduce code duplication.

The implementation in pkgbuild_parse.y is quite ugly, due to
the use of globals. However, without making Yacc re-entrant, it will have
to do.

Signed-off-by: Sebastian Nowicki <sebnow@gmail.com>